Windows: add NSIS installer builder
This commit adds NSIS installer support for Julia on Windows.

The NSIS installer is built uncompressed, compressed by 7-zip,
then combined into a self-extracting archive using the 7zS.sfx
module from the 7-zip extras distribution. (Thanks to Firefox
for this idea).

The default install directory is %LOCALAPPDATA%, but the user
is prompted to choose a directory before installation. After
installation, the install directory is opened in Windows
Explorer.